+++ development/application/api/http_root/cronjobs/__init__.py
@@ -4,6 +4,9 @@
 import cherrypy
 import common.http
 import common.tag_items
+import common.category_items
+import common.business_items
+import common.addresses
 
 
 @cherrypy.expose
@@ -16,3 +19,36 @@ def update_tag_items_cache():
     common.http.set_content_type_text()
     return u"OK"
 
+
+
+@cherrypy.expose
+def update_category_items_cache():
+
+    # Run job
+    common.category_items.update_category_items_cache()
+
+    # Finished
+    common.http.set_content_type_text()
+    return u"OK"
+
+
+@cherrypy.expose
+def update_business_items_cache():
+
+    # Run job
+    common.business_items.update_business_items_cache()
+
+    # Finished
+    common.http.set_content_type_text()
+    return u"OK"
+
+
+@cherrypy.expose
+def update_address_quantity_cache():
+
+    # Run job
+    common.addresses.update_address_quantity_cache()
+
+    # Finished
+    common.http.set_content_type_text()
+    return u"OK"